fix(dm-tool): split homebrew-items clone helper out of the IPC file
CI test runner can't load the `electron` native binary headlessly, so
importing `electron/ipc/homebrew-items.ts` (which `import { ipcMain }
from 'electron'`) crashed with "Electron failed to install correctly".

Move `stripIdentityForClone` + the `CompendiumItemTemplate` interface
to a sibling `homebrew-items-clone.ts` with no electron import. The
IPC file imports + re-exports them; tests and renderer code import
from the helper file directly.

Mirrors how `electron/ipc/items.test.ts` keeps clear of `items.ts` by
testing helpers from `compendium/image-url.ts` instead.
